backend/requirements.txt
aiofiles==23.2.1
aiosqlite==0.20.0
anyio==3.7.0
asgiref==3.8.1
asyncua==1.0.6
attrs==23.1.0
black>=24.8.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.6
Django==4.2.15
django-cors-headers==4.4.0
django-filter==23.2
djangorestframework==3.15.2
dulwich==0.21.5
exceptiongroup==1.1.1
fastapi==0.68.0
filelock==3.12.2
flake8==7.1.1
greenlet==2.0.2
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.2.3
keyring==23.13.1
lockfile==0.12.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.4
nest-asyncio==1.6.0
packaging==24.1
pathspec==0.12.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs>=2,<4.0.0
poetry>=1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
priority==2.0.0
protobuf==3.20.3
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.21
pydantic==1.8.2
pyflakes==3.2.0
PyMySQL==1.0.3
pyOpenSSL==24.0.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
pywin32-ctypes==0.2.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.0
sqlparse==0.5.1
starlette==0.14.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.16
uvicorn==0.15.0
virtualenv==20.23.0
webencodings==0.5.1
wsproto==1.2.0
zipp==3.15.0
